arreglando problemas en el dto, domain y entity, cambiando tipo de algunas variables

+++ src/main/java/com/sedikev/domain/model/AnimalDomain.java
@@ -2,13 +2,15 @@ package com.sedikev.domain.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class AnimalDomain {
 
     private String id;
     private LoteDomain loteDomain;
     private String nombre;
-    private Float peso;
+    private BigDecimal peso;
     private String sexo;
     private Integer numLote;
