+++ src/main/java/com/sedikev/domain/model/CarteraDomain.java
@@ -2,11 +2,13 @@ package com.sedikev.domain.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CarteraDomain {
 
     private Long id;
     private UsuarioDomain usuarioDomain;
-    private Float saldo;
+    private BigDecimal saldo;
 
 }
