+++ src/main/java/com/sedikev/domain/model/PagoDomain.java
@@ -2,6 +2,7 @@ package com.sedikev.domain.model;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -13,7 +14,7 @@ public class PagoDomain {
     private VentaDomain ventaDomain;
     private UsuarioDomain usuarioDomain;
     private String tipo_pago;
-    private Float cantidad;
+    private BigDecimal cantidad;
     private String descripcion;
     private Date fecha;
 }
